Agregando fetch y librerias

+++ js/carrito.js
@@ -66,6 +66,25 @@ function actualizarBotonesEliminar()  {
 }
 
 function eliminarDelCarrito(e) {
+    Toastify({
+        text: "ELIMINDADO DE CARRITO",
+        duration: 3500,
+        close: true,
+        gravity: "top", // `top` or `bottom`
+        position: "right", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+            background: "linear-gradient(to right, rgb(43, 52, 57), rgb(96, 189, 255))",
+            borderRadius: `1.5rem`,
+            fontSize: `0.8rem`,
+        },
+        offset: {
+            x: `1.5rem`, // horizontal axis - can be a number or a string indicating unity. eg: '2em'
+            y: `1.5rem`, // vertical axis - can be a number or a string indicating unity. eg: '2em'
+        },
+        onClick: function(){} // Callback after click
+    }).showToast();
+
     const idBoton = e.currentTarget.id;
     const index = productosEnCarrito.findIndex(producto => producto.id === idBoton);
     productosEnCarrito.splice(index, 1);
@@ -77,9 +96,23 @@ function eliminarDelCarrito(e) {
 botonVaciar.addEventListener("click", vaciarCarrito);
 
 function vaciarCarrito(){
-    productosEnCarrito.length = 0;
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
-    cargarProductosCarrito();
+    Swal.fire({
+        title: "¿Esta seguro que desea vaciar carrito?",
+        icon: "info",
+        html: `Se van a eliminar {productosEnCarrito.reduce((acc, producto) => acc + producto.cantidad, 0)} productos.`,
+        showCloseButton: true,
+        showCancelButton: true,
+        focusConfirm: false,
+        confirmButtonText: `SI`,
+        cancelButtonText: `NO`,
+    }).then((result) => {
+        /* Read more about isConfirmed, isDenied below */
+        if (result.isConfirmed) {
+            productosEnCarrito.length = 0;
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+            cargarProductosCarrito();
+        }
+    });
 }
 
 //Total Carrito
